+++ BakerSoft/Models/Tax.cs
@@ -42,9 +42,9 @@ namespace BakerSoft.Models
     class Tax
     {
         public int TaxId { get; set; }
-        public double SGST { get; set; }
-        public double CGST { get; set; }
-        public double TaxTotal
+        public decimal SGST { get; set; }
+        public decimal CGST { get; set; }
+        public decimal TaxTotal
         {
             get
             {
